pyproject.toml
[tool.poetry]
name = "bi-wealth"
version = "0.1.0"
description = ""
authors = ["Davi Henrique <davihscoelho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
duckdb = "^1.1.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
gspread = "^6.1.4"
google-auth = "^2.37.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
